# backend/docs/AssetFlow_OpenAPI_v1.0.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/employees":{"get":{"operationId":"EmployeeController_findAll","parameters":[],"responses":{"200":{"description":"Employees retrieved successfully"}},"security":[{"JWT-auth":[]}],"summary":"List all employee records (Admin & Manager)","tags":["Employees"]},"post":{"operationId":"EmployeeController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEmployeeDto"}}}},"responses":{"201":{"description":"Employee created successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Create a new employee record (Admin only)","tags":["Employees"]}},"/employees/{id}":{"get":{"operationId":"EmployeeController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Employee details returned"},"404":{"description":"Employee not found"}},"security":[{"JWT-auth":[]}],"summary":"Get employee details by ID (Admin & Manager)","tags":["Employees"]},"put":{"operationId":"EmployeeController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEmployeeDto"}}}},"responses":{"200":{"description":"Employee updated successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Update employee record (Admin only)","tags":["Employees"]},"delete":{"operationId":"EmployeeController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Employee deleted successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Delete an employee record (Admin only)","tags":["Employees"]}},"/assets":{"get":{"operationId":"AssetController_findAll","parameters":[],"responses":{"200":{"description":"Assets retrieved successfully"}},"security":[{"JWT-auth":[]}],"summary":"List all assets with optional filtering (Admin & Manager)","tags":["Assets"]},"post":{"operationId":"AssetController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssetDto"}}}},"responses":{"201":{"description":"Asset created successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Register a new asset (Admin only)","tags":["Assets"]}},"/assets/{id}/history":{"get":{"operationId":"AssetController_history","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"tags":["Assets"]}},"/assets/{id}/status":{"patch":{"operationId":"AssetController_changeStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetStatusDto"}}}},"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"tags":["Assets"]}},"/assets/{id}":{"get":{"operationId":"AssetController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Asset details returned"},"404":{"description":"Asset not found"}},"security":[{"JWT-auth":[]}],"summary":"Get asset details by ID (Admin & Manager)","tags":["Assets"]},"put":{"operationId":"AssetController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAssetDto"}}}},"responses":{"200":{"description":"Asset updated successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Update asset specifications (Admin only)","tags":["Assets"]},"patch":{"operationId":"AssetController_patch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAssetDto"}}}},"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"tags":["Assets"]},"delete":{"operationId":"AssetController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Asset retired successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Retire an asset and preserve history (Admin only)","tags":["Assets"]}},"/assets/{id}/assign/{employeeId}":{"patch":{"operationId":"AssetController_assignToEmployee","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"employeeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Assign asset directly to an employee (Admin & Manager)","tags":["Assets"]}},"/assets/{id}/unassign":{"patch":{"operationId":"AssetController_unassignFromEmployee","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"JWT-auth":[]}],"summary":"Unassign asset from employee (Admin & Manager)","tags":["Assets"]}},"/assignments":{"get":{"operationId":"AssignmentsController_findAll","parameters":[],"responses":{"200":{"description":"Assignments retrieved successfully"}},"security":[{"JWT-auth":[]}],"summary":"List asset assignments (Admin & Manager)","tags":["Assignments"]},"post":{"operationId":"AssignmentsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssignmentDto"}}}},"responses":{"201":{"description":"Asset assigned successfully"},"400":{"description":"Asset unavailable or employee inactive"}},"security":[{"JWT-auth":[]}],"summary":"Assign an available asset to an active employee (Admin & Manager)","tags":["Assignments"]}},"/assignments/{id}":{"get":{"operationId":"AssignmentsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Assignment details returned"},"404":{"description":"Assignment not found"}},"security":[{"JWT-auth":[]}],"summary":"Get assignment details by ID (Admin & Manager)","tags":["Assignments"]}},"/returns":{"post":{"operationId":"ReturnsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReturnDto"}}}},"responses":{"201":{"description":"Return processed and asset status updated"},"400":{"description":"Assignment already returned or invalid condition"},"404":{"description":"Assignment not found"}},"security":[{"JWT-auth":[]}],"summary":"Process equipment return and inspect condition (Admin & Manager)","tags":["Returns"]}},"/categories":{"get":{"operationId":"CategoriesController_findAll","parameters":[],"responses":{"200":{"description":"Categories retrieved successfully"}},"security":[{"JWT-auth":[]}],"summary":"List all asset categories (Admin & Manager)","tags":["Categories"]},"post":{"operationId":"CategoriesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"Category created successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Create a new asset category (Admin only)","tags":["Categories"]}},"/categories/{id}":{"get":{"operationId":"CategoriesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Category details returned"},"404":{"description":"Category not found"}},"security":[{"JWT-auth":[]}],"summary":"Get category details by ID (Admin & Manager)","tags":["Categories"]},"patch":{"operationId":"CategoriesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":"Category updated successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Update an asset category (Admin only)","tags":["Categories"]},"delete":{"operationId":"CategoriesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Category deleted successfully"},"403":{"description":"Forbidden - requires ADMIN role"}},"security":[{"JWT-auth":[]}],"summary":"Delete an asset category (Admin only)","tags":["Categories"]}},"/asset-history":{"get":{"operationId":"AssetHistoryController_findAll","parameters":[],"responses":{"200":{"description":"Asset history records retrieved successfully"}},"security":[{"JWT-auth":[]}],"summary":"List all asset history audit logs (Admin & Manager)","tags":["Asset History"]}},"/asset-history/asset/{assetId}":{"get":{"operationId":"AssetHistoryController_findByAsset","parameters":[{"name":"assetId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Asset lifecycle events returned"}},"security":[{"JWT-auth":[]}],"summary":"Get lifecycle audit trail for a specific asset (Admin & Manager)","tags":["Asset History"]}},"/dashboard/summary":{"get":{"operationId":"DashboardController_getStatistics","parameters":[],"responses":{"200":{"description":"Dashboard metrics and summaries retrieved successfully"}},"security":[{"JWT-auth":[]}],"summary":"Get KPI statistics, status distribution, and recent activity (Admin & Manager)","tags":["Dashboard"]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"Health check completed successfully"}},"summary":"Check API and database health connectivity","tags":["Health"]}},"/employee-portal/dashboard":{"get":{"operationId":"EmployeePortalController_getMyDashboard","parameters":[],"responses":{"200":{"description":"Employee dashboard metrics and assigned assets returned"}},"security":[{"JWT-auth":[]}],"summary":"Get dashboard overview for the authenticated employee","tags":["Employee Portal"]}},"/employee-portal/profile":{"get":{"operationId":"EmployeePortalController_getMyProfile","parameters":[],"responses":{"200":{"description":"Employee profile returned"}},"security":[{"JWT-auth":[]}],"summary":"Get profile information for the authenticated employee","tags":["Employee Portal"]}},"/employee-portal/assets":{"get":{"operationId":"EmployeePortalController_getMyAssets","parameters":[],"responses":{"200":{"description":"List of assigned assets returned"}},"security":[{"JWT-auth":[]}],"summary":"Get assets currently assigned to the authenticated employee","tags":["Employee Portal"]}},"/employee-portal/assignments":{"get":{"operationId":"EmployeePortalController_getMyAssignments","parameters":[],"responses":{"200":{"description":"List of employee assignments returned"}},"security":[{"JWT-auth":[]}],"summary":"Get assignment history records for the authenticated employee","tags":["Employee Portal"]}},"/employee-portal/history":{"get":{"operationId":"EmployeePortalController_getMyHistory","parameters":[],"responses":{"200":{"description":"List of employee lifecycle history events returned"}},"security":[{"JWT-auth":[]}],"summary":"Get lifecycle event history involving the authenticated employee","tags":["Employee Portal"]}},"/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"List of system users"}},"security":[{"JWT-auth":[]}],"summary":"List all system users (Admin only)","tags":["Users"]},"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User created successfully"}},"security":[{"JWT-auth":[]}],"summary":"Create a new system user account (Admin only)","tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"User details"}},"security":[{"JWT-auth":[]}],"summary":"Get user details by ID (Admin only)","tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"User updated successfully"}},"security":[{"JWT-auth":[]}],"summary":"Update user position, role, or active status (Admin only)","tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"User deleted successfully"}},"security":[{"JWT-auth":[]}],"summary":"Delete a user account (Admin only)","tags":["Users"]}},"/users/{id}/status":{"patch":{"operationId":"UsersController_toggleStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"User status toggled"}},"security":[{"JWT-auth":[]}],"summary":"Toggle user active/deactive status (Admin only)","tags":["Users"]}}},"info":{"title":"AssetFlow API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{}},"CreateEmployeeDto":{"type":"object","properties":{"employeeNo":{"type":"string","example":"EMP-001","description":"Unique company employee ID / badge number"},"firstName":{"type":"string","example":"Alice","description":"First name"},"lastName":{"type":"string","example":"Johnson","description":"Last name"},"email":{"type":"string","example":"alice.johnson@company.com","description":"Corporate email address"},"department":{"type":"string","example":"Engineering","description":"Department name"},"position":{"type":"string","example":"Lead Software Architect","description":"Job position / designation"},"isActive":{"type":"boolean","default":true,"description":"Whether the employee is active and eligible to receive equipment"}},"required":["employeeNo","firstName","lastName","email","department","position"]},"UpdateEmployeeDto":{"type":"object","properties":{"employeeNo":{"type":"string","example":"EMP-001","description":"Unique company employee ID / badge number"},"firstName":{"type":"string","example":"Alice","description":"First name"},"lastName":{"type":"string","example":"Johnson","description":"Last name"},"email":{"type":"string","example":"alice.johnson@company.com","description":"Corporate email address"},"department":{"type":"string","example":"Engineering","description":"Department name"},"position":{"type":"string","example":"Principal Software Engineer","description":"Job position / designation"},"isActive":{"type":"boolean","example":true,"description":"Active employment status"}}},"AssetStatusDto":{"type":"object","properties":{}},"CreateAssetDto":{"type":"object","properties":{"assetTag":{"type":"string","example":"LAP-0012","description":"Unique asset inventory tag"},"name":{"type":"string","example":"Dell Latitude 5450","description":"Equipment name/title"},"category":{"type":"string","example":"Laptops","description":"Asset category name"},"brand":{"type":"string","example":"Dell","description":"Brand or manufacturer"},"model":{"type":"string","example":"Latitude 5450","description":"Hardware model"},"serialNumber":{"type":"string","example":"SN-5450-9988","description":"Manufacturer serial number"},"status":{"type":"string","enum":["available","assigned","damaged","under_repair","lost","retired"],"default":"available","description":"Current lifecycle status of the asset"},"purchaseDate":{"type":"string","example":"2024-01-15","description":"Purchase date in YYYY-MM-DD format"},"purchasePrice":{"type":"number","example":1299.99,"description":"Purchase price in USD"},"warrantyExpiryDate":{"type":"string","example":"2026-12-31","description":"Warranty expiry date in YYYY-MM-DD format"}},"required":["assetTag","name","category"]},"UpdateAssetDto":{"type":"object","properties":{}},"CreateAssignmentDto":{"type":"object","properties":{}},"CreateReturnDto":{"type":"object","properties":{"assignmentId":{"type":"number","example":1,"description":"ID of the ACTIVE assignment being returned"},"condition":{"type":"string","example":"GOOD","enum":["NEW","GOOD","FAIR","DAMAGED"],"description":"Hardware condition upon return (GOOD/NEW/FAIR returns to available; DAMAGED sets status to damaged)"},"notes":{"type":"string","example":"Equipment returned in pristine shape with charger","description":"Inspection or return notes"}},"required":["assignmentId","condition"]},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string","example":"Laptops","description":"Name of the equipment category"},"description":{"type":"string","example":"Enterprise notebooks and ultrabooks","description":"Detailed category description"},"active":{"type":"boolean","default":true,"description":"Whether the category is active for equipment assignment"}},"required":["name"]},"UpdateCategoryDto":{"type":"object","properties":{"name":{"type":"string","example":"Laptops & Workstations","description":"Updated category name"},"description":{"type":"string","example":"Updated category description","description":"Updated category description"},"active":{"type":"boolean","example":false,"description":"Active status of the category"}}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"sarah.connor@assetflow.com","description":"Work email address"},"password":{"type":"string","example":"UserPassword123!","description":"Account password (defaults to User123! if omitted)"},"firstName":{"type":"string","example":"Sarah","description":"First name"},"lastName":{"type":"string","example":"Connor","description":"Last name"},"position":{"type":"string","example":"IT Support Specialist","description":"User job position / title"},"role":{"type":"string","enum":["ADMIN","MANAGER","EMPLOYEE"],"default":"EMPLOYEE","description":"System role"},"isActive":{"type":"boolean","default":true,"description":"Active account status"}},"required":["email","firstName","lastName"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"sarah.connor@assetflow.com","description":"Work email address"},"password":{"type":"string","example":"NewPassword123!","description":"Optional new password"},"firstName":{"type":"string","example":"Sarah","description":"First name"},"lastName":{"type":"string","example":"Connor","description":"Last name"},"position":{"type":"string","example":"Operations Lead","description":"User job position / title"},"role":{"type":"string","enum":["ADMIN","MANAGER","EMPLOYEE"],"description":"System role"},"isActive":{"type":"boolean","description":"Active account status"}}}}}}